+++ Control/Joint/Effects.hs
@@ -1,5 +1,6 @@
 module Control.Joint.Effects (module Exports) where
 
+import Control.Joint.Effects.List as Exports
 import Control.Joint.Effects.Maybe as Exports
 import Control.Joint.Effects.Either as Exports
 import Control.Joint.Effects.Store as Exports